combine subprocess_utils.hpp and subprocess.hpp into the same header
